feat(TF-52): 내 기록 페이지 next.js 마이그레이션
